LayoutTests/imported/w3c/web-platform-tests/css/css-display/accessibility/display-contents-role-and-label-expected.txt

Testing accessibility of display: contents
Verifies accessibility behavior of the contents keyword for the CSS display property

Accname
hidden label visible to all users label x x x x label x label label label
Accessibility roles
x x x x
x	x
x	x
x	x
x	x
x	x
x	x
x
x
x
x x x x x x x x x x x x x x x x x x x x x x x x x
y
x
y
x
y
x
y
x
y
x
y
x
y
x
y
x
y
x	x
x	x
x
x
x
x
x	x
x	x
x
x
x
x	x
x	x
x
x

PASS tr with display: contents, within table with display: flex, has row role
PASS td as child of tr with display: contents, within table with display: flex, has cell role
PASS tr with display: contents, within table with role=table with display: flex, has row role
PASS td as child of tr with display: contents, within table with role=table with display: flex, has cell role
PASS tr with display: contents, within table with role=grid with display: flex, has row role
PASS td (no explicit role) as child of tr with display: contents, within table with role=grid with display: flex, has gridcell role
PASS ul with display: contents, as child of div with display: grid, has list role
PASS listitem within ul with display: contents, as child of div with display: grid, has listitem role
PASS header with display: contents has banner role
PASS main with display: contents has main role
PASS footer with display: contents has contentinfo role
PASS search with display: contents has search role
PASS div with role banner and display: contents has banner role
PASS div with role main and display: contents has main role
PASS div with role contentinfo and display: contents has contentinfo role
PASS ul with role list and display: contents (child li has display: contents) has list role
PASS li as child of ul with role list, both with display: contents, has listitem role
PASS ul with role list and display: contents has list role
PASS li, as child of ul with role list and display: contents, has listitem role
PASS li with display: contents, as child of ul with role list, has listitem role
PASS ol with role list and display: contents has list role (child li has display: contents)
PASS li as child of ol with role list, both with display: contents, has listitem role
PASS ol with role list and display: contents has list role
PASS li, as child of ol with role list and display: contents, has listitem role
PASS li with display: contents, as child of ol with role list, has listitem role
PASS div with list role and display: contents has list role
PASS div with list role and display: contents has list role (child div with listitem role has display: contents)
PASS table with role table and display: contents has table role
PASS tr in table with role table and display: contents has row role
PASS td within tr in table with role table, all with display: contents, has cell role
PASS div with role table with display: contents has table role
PASS div with role rowgroup in div with role table, both with display: contents, has rowgroup role
PASS td within table with role grid, both with display: contents, has gridcell role
PASS div with role grid with display: contents has grid role
PASS div with role rowgroup in div with role grid, both with display: contents, has rowgroup role
PASS table with role treegrid and display: contents has treegrid role
PASS div with role treegrid and display: contents has treegrid role
PASS ul with role tree and display: contents has tree role
PASS li with role treeitem and display: contents, within ul with role tree and display: contents, has treeitem role
PASS li with role treeitem and display: contents, within ul with role=group (within ul with role tree and display: contents), has treeitem role
PASS Label: button with display: contents is labelled via referenced hidden span
PASS Label: link with display: contents is labelled via aria-labelledby
PASS Label: nav with display: contents is labelled via aria-labelledby
PASS Label: button with display: contents is labelled via aria-label
PASS Label: div with role group with display: contents is labelled via aria-label
PASS Label: button with display: contents is labelled via label for/id association
PASS Label: g element with display: contents, as child of svg, is labelled via title element
PASS Label: link with display: contents is labelled via name from contents
PASS Label: button with display: contents is labelled via name from contents
PASS Label: h3 with display: contents is labelled via name from contents
PASS Label: button with display: contents has button role
PASS Label: button with explicit button role and display: contents has button role
PASS Label: div with button role and display: contents has button role
PASS Label: div with button role, tabindex=0 and display: contents has button role
FAIL Label: th as child of tr with display: contents, within table with display: flex, has columnheader role assert_equals: <th data-expectedrole="columnheader" data-expectedlabel="x" data-testname="th as child of tr with display: contents, within table with display: flex, has columnheader role" class="ex-role-and-label ex-label-only ex-role-only">x</th> expected "x" but got ""
FAIL Label: th as child of tr with display: contents, within table with role=table with display: flex, has columnheader role assert_equals: <th data-expectedrole="columnheader" data-expectedlabel="x" data-testname="th as child of tr with display: contents, within table with role=table with display: flex, has columnheader role" class="ex-role-and-label ex-label-only ex-role-only">x</th> expected "x" but got ""
FAIL Label: th as child of tr with display: contents, within table with role=grid with display: flex, has columnheader role assert_equals: <th data-expectedrole="columnheader" data-expectedlabel="x" data-testname="th as child of tr with display: contents, within table with role=grid with display: flex, has columnheader role" class="ex-role-and-label ex-label-only ex-role-only">x</th> expected "x" but got ""
PASS Label: h3 with display: contents has heading role
PASS Label: div with heading role, aria-level=3 and display: contents has heading role
PASS Label: span with heading role, aria-level=1 and display: contents has heading role
PASS Label: nav with display: contents and aria-label has navigation role
PASS Label: aside with display: contents and aria-label has complementary role
PASS Label: form with display: contents has form role
PASS Label: section with aria-label and display: contents has region role
PASS Label: div with role navigation, aria-label and display: contents has navigation role
PASS Label: div with role complementary, aria-label and display: contents has complementary role
PASS Label: div with role form, aria-label and display: contents has form role
PASS Label: div with role search and display: contents has search role
PASS Label: div with role region, aria-label and display: contents has region role
PASS Label: link with display: contents has link role
PASS Label: link with explicit link role and display: contents has link role
PASS Label: div with link role and display: contents has link role
PASS Label: div with link role, tabindex=0 and display: contents has link role
FAIL Label: div with listitem role, as child of div with display: contents, has listitem role assert_equals: <div role="listitem" data-expectedrole="listitem" data-expectedlabel="x" data-testname="div with listitem role, as child of div with display: contents, has listitem role" class="ex-role-and-label ex-label-only ex-role-only">x</div> expected "x" but got ""
FAIL Label: div with listitem role (as child of div with list role), both with display: contents, has listitem role assert_equals: <div role="listitem" style="display: contents;" data-expectedrole="listitem" data-expectedlabel="y" data-testname="div with listitem role (as child of div with list role), both with display: contents, has listitem role" class="ex-role-and-label ex-label-only ex-role-only">y</div> expected "y" but got ""
FAIL Label: div with listitem role with display: contents, as child of div with list role, has listitem role assert_equals: <div role="listitem" style="display: contents;" data-expectedrole="listitem" data-expectedlabel="y" data-testname="div with listitem role with display: contents, as child of div with list role, has listitem role" class="ex-role-and-label ex-label-only ex-role-only">y</div> expected "y" but got ""
FAIL Label: td within tr both with display: contents, in table with role table and display: contents, has columnheader role assert_equals: <th style="display: contents;" data-expectedrole="columnheader" data-expectedlabel="x" data-testname="td within tr both with display: contents, in table with role table and display: contents, has columnheader role" class="ex-role-and-label ex-label-only ex-role-only">x</th> expected "x" but got ""
FAIL Label: div with role columnheader with display: contents, in div with role rowgroup within div with table role both with display: contents, has columnheader role assert_equals: <div role="columnheader" style="display: contents;" data-expectedrole="columnheader" data-expectedlabel="x" data-testname="div with role columnheader with display: contents, in div with role rowgroup within div with table role both with display: contents, has columnheader role" class="ex-role-and-label ex-label-only ex-role-only">x</div> expected "x" but got ""
FAIL Label: th within table with role grid, both with display: contents, has columnheader role assert_equals: <th style="display: contents;" data-expectedrole="columnheader" data-expectedlabel="x" data-testname="th within table with role grid, both with display: contents, has columnheader role" class="ex-role-and-label ex-label-only ex-role-only">x</th> expected "x" but got ""
FAIL Label: div with role columnheader with display: contents, in div with role rowgroup within div with grid role both with display: contents, has rowgroup role assert_equals: <div role="columnheader" style="display: contents;" data-expectedrole="columnheader" data-expectedlabel="x" data-testname="div with role columnheader with display: contents, in div with role rowgroup within div with grid role both with display: contents, has rowgroup role" class="ex-role-and-label ex-label-only ex-role-only">x</div> expected "x" but got ""
FAIL Label: div with gridcell role and display: contents, within div with role row contained in a div with role grid both with display: contents, has gridcell role assert_equals: <div role="gridcell" style="display: contents;" data-expectedrole="gridcell" data-expectedlabel="x" data-testname="div with gridcell role and display: contents, within div with role row contained in a div with role grid both with display: contents, has gridcell role" class="ex-role-and-label ex-label-only ex-role-only">x</div> expected "x" but got ""
PASS Role: button with display: contents is labelled via referenced hidden span
PASS Role: link with display: contents is labelled via aria-labelledby
PASS Role: nav with display: contents is labelled via aria-labelledby
PASS Role: button with display: contents is labelled via aria-label
PASS Role: div with role group with display: contents is labelled via aria-label
PASS Role: button with display: contents is labelled via label for/id association
PASS Role: g element with display: contents, as child of svg, is labelled via title element
PASS Role: link with display: contents is labelled via name from contents
PASS Role: button with display: contents is labelled via name from contents
PASS Role: h3 with display: contents is labelled via name from contents
PASS Role: button with display: contents has button role
PASS Role: button with explicit button role and display: contents has button role
PASS Role: div with button role and display: contents has button role
PASS Role: div with button role, tabindex=0 and display: contents has button role
PASS Role: th as child of tr with display: contents, within table with display: flex, has columnheader role
PASS Role: th as child of tr with display: contents, within table with role=table with display: flex, has columnheader role
PASS Role: th as child of tr with display: contents, within table with role=grid with display: flex, has columnheader role
PASS Role: h3 with display: contents has heading role
PASS Role: div with heading role, aria-level=3 and display: contents has heading role
PASS Role: span with heading role, aria-level=1 and display: contents has heading role
PASS Role: nav with display: contents and aria-label has navigation role
PASS Role: aside with display: contents and aria-label has complementary role
PASS Role: form with display: contents has form role
PASS Role: section with aria-label and display: contents has region role
PASS Role: div with role navigation, aria-label and display: contents has navigation role
PASS Role: div with role complementary, aria-label and display: contents has complementary role
PASS Role: div with role form, aria-label and display: contents has form role
PASS Role: div with role search and display: contents has search role
PASS Role: div with role region, aria-label and display: contents has region role
PASS Role: link with display: contents has link role
PASS Role: link with explicit link role and display: contents has link role
PASS Role: div with link role and display: contents has link role
PASS Role: div with link role, tabindex=0 and display: contents has link role
PASS Role: div with listitem role, as child of div with display: contents, has listitem role
PASS Role: div with listitem role (as child of div with list role), both with display: contents, has listitem role
PASS Role: div with listitem role with display: contents, as child of div with list role, has listitem role
PASS Role: td within tr both with display: contents, in table with role table and display: contents, has columnheader role
PASS Role: div with role columnheader with display: contents, in div with role rowgroup within div with table role both with display: contents, has columnheader role
PASS Role: th within table with role grid, both with display: contents, has columnheader role
PASS Role: div with role columnheader with display: contents, in div with role rowgroup within div with grid role both with display: contents, has rowgroup role
PASS Role: div with gridcell role and display: contents, within div with role row contained in a div with role grid both with display: contents, has gridcell role
PASS div with display: contents has generic role
PASS span with display: contents has generic role
